\name{moons}

\alias{moons}
\docType{data}
\title{Moons of the 13 planets of the Solar System}

\description{
This dataset contains informations about dimensions, number of moons and distance from the Sun of the eight planets and the five dwarf planets of the Solar System.
}

\usage{data("moons")}
\format{
  A data frame with 13 observations on the following 5 variables.
  \describe{
    \item{\code{Name}}{name of the planet.}
    \item{\code{Distance}}{distance from the sun.}
    \item{\code{Diameter}}{diameter of the planet.}
    \item{\code{Mass}}{mass of the planet.}
    \item{\code{Moons}}{number of moons.}
  }
}

