\name{moons}

\alias{moons}

\title{Moons of the 13 planets of the Solar System}

\description{
This dataset contains informations about dimensions, number of moons and distance from the Sun of the eight planets and the five dwarf planets of the Solar System.
}
\usage{
data(moons, package = 'glmdiag')
}

\format{
A data frame with 13 observations on the following 5 variables:
\itemize{
\item{\emph{Name}: name of the planet}
\item{\emph{Distance}: distance from the sun}
\item{\emph{Diameter}: diameter of the planet}
\item{\emph{Mass}: mass of the planet}
\item{\emph{Moons}: number of moons}
}
}

\source{
???
}
